+++ GRDB/Record/TableRecord.swift
@@ -130,7 +130,7 @@ extension TableRecord {
     public static func selectionSQL(alias: String? = nil) -> String {
         let alias = TableAlias(tableName: databaseTableName, userName: alias)
         let selection = databaseSelection.map { $0.qualifiedSelectable(with: alias) }
-        var context = SQLGenerationContext.recordSelectionGenerationContext()
+        var context = SQLGenerationContext.selectionContext
         return selection
             .map { $0.resultColumnSQL(&context) }
             .joined(separator: ", ")